==> Kode/1_Preprocessing_Balonggandu_2022.ipynb <==
import pandas as pd
from datetime import datetime, timedelta

# Membuat dataframe kosong dengan kolom yang sama
data_columns = ["WAKTU", "TAHUN", "BULAN", "PEKAN", "TANGGAL", "HARI", "JAM", 
                "MOTOR", "MOBIL", "TRUK/BUS", "JUMLAH", "SMP_MTR", "SMP_MBL", "SMP_TRK", "SMP"]


# Membuat rentang waktu untuk satu tahun, dimulai dari 1 Januari 2022
start_date = datetime(2022, 1, 1, 0, 0)
end_date = datetime(2022, 12, 31, 23, 45)
date_range = pd.date_range(start=start_date, end=end_date, freq='15T')

# Membuat dataframe dengan tanggal tersebut
data_empty = pd.DataFrame(date_range, columns=["WAKTU"])
data_empty['TAHUN'] = data_empty['WAKTU'].dt.year
data_empty['BULAN'] = data_empty['WAKTU'].dt.month
data_empty['PEKAN'] = data_empty['WAKTU'].apply(lambda x: (x.day - 1) // 7 + 1)
data_empty['TANGGAL'] = data_empty['WAKTU'].dt.day
data_empty['HARI'] = data_empty['WAKTU'].dt.day_name()
data_empty['JAM'] = data_empty['WAKTU'].dt.strftime('%H:%M')

# Menambahkan kolom data yang kosong
for col in ["MOTOR", "MOBIL", "TRUK/BUS", "JUMLAH", "SMP_MTR", "SMP_MBL", "SMP_TRK", "SMP"]:
    data_empty[col] = None

data_empty.head()

data_empty.to_excel('../Data/Preprocessing/Data Balonggandu/2022/1_missing_times_balonggandu_2022.xlsx', index=False)
