==> Kode/3_Check_Losarang_2023.ipynb <==
import os
import numpy as np
import pandas as pd

def read_excel(path):
    return pd.read_excel(path)

def menggabungkan_data(df, missing_times):
    # Menggabungkan data hilang dengan DataFrame utama
    df = pd.concat([df, missing_times], ignore_index=True)
    
    # Mengurutkan DataFrame berdasarkan kolom 'Waktu'
    df = df.sort_values(by=['TAHUN','BULAN','TANGGAL','HARI','JAM']).reset_index(drop=True)

    return df

path_1 = '../Data/Preprocessing/Data Losarang/2023/3_clean_losarang_2023.xlsx'
path_2 = '../Data/Preprocessing/Data Losarang/2023/3_outliers_losarang_2023.xlsx'

df_clean = read_excel(path_1)
df_outlier = read_excel(path_2)

df_compiled = menggabungkan_data(df_clean, df_outlier)
df_compiled.head()

len(df_compiled)

# Buat daftar waktu yang lengkap dari 2022-01-01 00:00:00 hingga 2022-12-31 23:45:00 dengan interval 15 menit
time_range = pd.date_range(start='2023-01-01 00:00:00', end='2023-12-31 23:45:00', freq='15T')

time_range

missing_time = set(time_range) - set(df_compiled['WAKTU'])
print("Waktu yang hilang:", missing_time)

Waktu yang hilang: {Timestamp('2023-11-01 00:00:00'), Timestamp('2023-09-27 00:00:00'), Timestamp('2023-09-25 00:00:00'), Timestamp('2023-10-08 00:00:00'), Timestamp('2023-10-19 00:00:00'), Timestamp('2023-10-07 00:00:00'), Timestamp('2023-09-24 00:00:00'), Timestamp('2023-02-01 00:00:00'), Timestamp('2023-12-01 00:00:00'), Timestamp('2023-09-26 00:00:00')}
