==> Kode/3_LHRT_Balonggandu_2023.ipynb <==
import os
import numpy as np
import pandas as pd

def read_excel(path):
    return pd.read_excel(path)

def delete_outlier(df, HARI, JAM):
    df = df[(df['HARI'] == HARI) & (df['JAM'] == JAM)]    
    
    # Menghitung Q1 (kuartil pertama) dan Q3 (kuartil ketiga)
    Q1 = df['SMP'].quantile(0.25)
    Q3 = df['SMP'].quantile(0.75)
    IQR = Q3 - Q1
    
    # Menentukan batas bawah dan batas atas untuk outlier
    lower_bound = Q1 - 1.5 * IQR
    upper_bound = Q3 + 1.5 * IQR
    
    # Menemukan dan menghapus outlier
    df_outlier = df[(df['SMP'] < lower_bound) | (df['SMP'] > upper_bound)]
    df_cleaned = df[(df['SMP'] >= lower_bound) & (df['SMP'] <= upper_bound)]
    
    # if(len(df_outlier)>0):
    #     print('Ada Outlier')

    # df_grouped = df_cleaned[['HARI', 'JAM', 'SMP']]
    # df_grouped = round(df_grouped.groupby(['HARI', 'JAM']).mean()).reset_index()

    return df_outlier, df_cleaned


path_1 = '../Data/Preprocessing/Data Balonggandu/2023/2_input_balonggandu_2023.xlsx'

df = read_excel(path_1)
df.head()

df_waktu = df[['HARI','JAM']]
df_waktu = df_waktu.groupby(['HARI', 'JAM']).mean().reset_index()
df_waktu.head()

all_df_outlier = []
all_df_cleaned = []

for i, row in df_waktu.iterrows():
    df_outlier,  df_cleaned = delete_outlier(df, row['HARI'], row['JAM'])
    all_df_outlier.append(df_outlier)
    all_df_cleaned.append(df_cleaned)

# gabungkan data
df_outlier_combined = pd.concat(all_df_outlier)
df_cleaned_combined = pd.concat(all_df_cleaned)

df_outlier_combined.head()

df_cleaned_combined.head()

df_outlier_combined = df_outlier_combined.sort_values(by='WAKTU').reset_index(drop=True)
df_cleaned_combined = df_cleaned_combined.sort_values(by='WAKTU').reset_index(drop=True)

df_cleaned_combined.to_excel('../Data/Preprocessing/Data Balonggandu/2023/3_clean_balonggandu_2023.xlsx', index=False)
df_outlier_combined.to_excel('../Data/Preprocessing/Data Balonggandu/2023/3_outliers_balonggandu_2023.xlsx', index=False)

df_cleaned_combined.head()

df_smp = df_cleaned_combined[['WAKTU','TAHUN', 'BULAN', 'PEKAN' ,'TANGGAL', 'HARI','JAM','MOTOR','MOBIL','TRUK/BUS',
                              'SMP_MTR','SMP_MBL','SMP_TRK','SMP']]

df_smp.head(8)

# Round down the WAKTU column to the nearest hour
df_smp_jam = df_smp.copy()
df_smp_jam['JAM'] = df_smp['WAKTU'].dt.floor('H')

# Drop the original WAKTU column if it's not needed anymore
df_smp_jam = df_smp_jam.drop(columns=['WAKTU'])

# Format the 'JAM' column to show only hour and minute
df_smp_jam['JAM'] = df_smp_jam['JAM'].dt.strftime('%H:%M')

df_smp_jam.head(8)

# Group by the rounded-down hour and sum the SMP column
df_smp_jam = df_smp_jam.groupby(['TAHUN', 'BULAN', 'PEKAN', 'TANGGAL', 'HARI', 'JAM'], as_index=False).mean().round()

df_smp_jam.head(8)

# List kolom yang akan dikalikan 4
columns_to_multiply = ['MOTOR', 'MOBIL', 'TRUK/BUS', 'SMP_MTR', 'SMP_MBL', 'SMP_TRK', 'SMP']

# Mengalikan nilai kolom dengan 4
df_smp_jam[columns_to_multiply] = df_smp_jam[columns_to_multiply] * 4

df_smp_jam.head(8)

df_lhrt = df_smp_jam.copy()

df_lhrt = df_lhrt.drop(columns=['TAHUN','BULAN','TANGGAL','PEKAN'])

df_lhrt.head()

# df_lhrt  = df_lhrt.groupby(['PEKAN', 'HARI', 'JAM'], as_index=False).mean().round()
df_lhrt  = df_lhrt.groupby(['HARI', 'JAM'], as_index=False).mean().round()

df_lhrt.head()

df_lhrt.to_excel('../Data/Preprocessing/Data Balonggandu/2023/3_lhrt_balonggandu_2023.xlsx', index=False)

len(df_lhrt)
