==> Kode/5_Koefisien_Balonggandu_2023.ipynb <==
import os
import numpy as np
import pandas as pd

def read_excel(path):
    return pd.read_excel(path)

def koefisien(df):
    data.loc[:, 'KOEFISIEN'] = data['SMP'] / data['SMP'].max()
    return data
    

path_1 = '../Data/Preprocessing/Data Balonggandu/2023/4_lhrt_normalisasi_balonggandu_2023.xlsx'
df = read_excel(path_1)

df['KOEFISIEN'] = df['SMP'] / df['SMP'].max()

df.head()

# Daftar kolom yang diinginkan dalam urutan baru
new_column_order = [
    'HARI', 'JAM', 'KOEFISIEN',
    'MOTOR', 'MOBIL', 'TRUK/BUS',
    'SMP_MTR','SMP_MBL','SMP_TRK',
    'SMP'
]
# Menyusun ulang kolom berdasarkan urutan yang diinginkan
df = df.reindex(columns=new_column_order)

df['SMP'] = np.nan
df['MOTOR'] = np.nan
df['MOBIL'] = np.nan
df['TRUK/BUS'] = np.nan
df['SMP_MTR'] = np.nan
df['SMP_MBL'] = np.nan
df['SMP_TRK'] = np.nan

df.head()

df.to_excel('../Data/Preprocessing/Data Balonggandu/2023/5_lhrt_koefisien_balonggandu_2023.xlsx', index=False)
