==> Kode/7_Analisis_Balonggandu_22_23.ipynb <==
import os
import numpy as np
import pandas as pd
import pickle

def read_excel(path):
    return pd.read_excel(path)

path_1 = '../Data/Preprocessing/Data Balonggandu/2023/2_input_balonggandu_2023.xlsx'
path_2 = '../Data/Preprocessing/Data Balonggandu/2022/2_input_balonggandu_2022.xlsx'

df_2023 = read_excel(path_1)
df_2022 = read_excel(path_2)

df_2023.head()

df_2022.head()

desc_2023 = df_2023['SMP'].describe()
desc_2022 = df_2022['SMP'].describe()

# Combine the statistics for better comparison
desc_combined = pd.DataFrame({
    '2023': desc_2023,
    '2022': desc_2022
})

desc_combined

sum_2023 = df_2023['SMP'].sum()

print(sum_2023)

print(13881165/13023570)

print(13023570/13881165)
